// src/api/generated/models/InventoryItemUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Update inventory item
 */
export type InventoryItemUpdate = {
    name?: (string | null);
    description?: (string | null);
    sku?: (string | null);
    unit?: (string | null);
    current_stock?: (number | string | null);
    min_stock?: (number | string | null);
    max_stock?: (number | string | null);
    unit_cost?: (number | string | null);
    supplier_id?: (string | null);
    location_id?: (string | null);
    category?: (string | null);
    is_tracked?: (boolean | null);
};